Added config file loader
By overloading the << operator, the program can now save all the solar system data to a file. When the program starts, it will generate a solar system based on the data file, if there is one in the same directory as Tutorial. Otherwise it will create a new solar system.

Also, the solar system saves when you exit the program by pressing Esc.